requirements.txt
requests>=2.32
python-dotenv>=1.0
rich>=15.0
